- Fixes GUID casing in csproj. - Update samples
* Project reference GUIDs in csproj are in lowercase to match the project GUID.
VS is inconsistent about the casing of GUID. When it defines a project GUID it's in uppercase, and when it's referencing another project it's in lowercase. In .sln it's in uppercase. Now the generation matches VS's convention.

(cherry picked from commit 7b7299536e39d3f3a3472b37e3a782ce2a51037b)