Project build dependency for solutions that don't have fastbuild projects. * Added a way to add a project build dependency in Visual Studio's solution files. * The new DependencySetting.BuildPrerequisite gathers a list of project's GUIDs, which is how Visual Studio resolves build dependencies. * The list of GUIDs is then applied in the VisualStudio's Sln generator only if the solution is not a fast build one, to preserve current behavior intact.
I wasn't too sure whether or not to add new Project.Configuration.AddBuildDependency method or to simply add a flag to DependencySetting. In the end I chose the later as I felt that it was easier for existing Sharpmake scripts to simple update their calls to AddPublicDependency/AddPrivateDependency with the additional flag.

Moreover, going through AddPublicDependency/AddPrivateDependency, circular dependencies are already found and throws an exception, so it's not possible to generate a solution that contains projects that requires themselves to be built first.

Tests: ran unit tests + tested with a Sharpmake script that generates a bunch of projects that need such build dependencies. Verified that the output solution in Vs2015 & Vs2017 contains the proper dependencies.