Fix library names that were shortened even in cases where they cannot be
Now we test both the command line prefix and the output lib prefix, and
only shorten the library name if it was prefixed by one or the other.

Warning: potentially the extension could be different from what the
platform expects by default by the shortened lib name, so we might need to
test for those as well.