Stubless DLL do not use target copy files anymore
If an EXE depends on a DLL, a custom build is added to copy the dll in the same directory as the EXE. This behavior is useful as the stub lib is auto-loading the DLL, so needs to find it in windows DLL searched path, and the EXE path is the prefered one. If the DLL has no stub libbed, it means that it will be explicitely loaded, so copying it is innapropriate.