Fix dynamic library dependency when you change OutputExtension option
OutputExtension is an option that change the default extension of the outputed binary of a project but this option is not well propagated as dependency:
When the final project (that generate executable) wants to copy its dependencies to the output directory, it use the hardcoded default extension.
I therefore use the correct extension in the dependency resolution which is in any case initialized with the default hardcoded value at the beginning, which should not change anything in nominal operation.