[XCode] Fix output and intermediate dirs
$OBJROOT is changed juste before building by XCode, so $CONFIGURATION_TEMP_DIR  was not the one wanted.
The reason is that some values like $OBJROOT or $SYMROOT are project specific and not target specific (according to https://pewpewthespells.com/blog/buildsettings.html#objroot). So XCode use the value defined in the project configs and not the target configs.
To set output paths, intermediate paths, per target and configuration we need to directly set the dedicated variable, like $CONFIGURATION_TEMP_DIR for example

Delete some redundancy