Resolve AdditionalSourceRootPaths before use
The AdditionalSourceRootPaths member never had the paths it contained resolved, resulting in paths that had mixed folder separators to not be looked up correctly inside of ResolveSourceFiles.

Ex: A path added to AdditionalSourcerootPaths on Mac OSX set as:

@"[project.SharpmakeCsPath]\..\externals\SomeExternal"

that expanded to:

/Users/SomeUser/Documents/SomeProject/sharpmake\\..\\externals\\SomeExternal

would fail to have the files in the directory looked up. Calling ResolvePath on the member causes the path to be normalized to:

/Users/SomeUser/Documents/SomeProject/externals/SomeExternal