AdditionalSourcePaths use case in HelloXCode sample
Enhance the HelloXCode sample to leverage the AdditionalSourcePaths project attribute. This helps to validate the previous bug fix for it. Not having the bugfix causes a compiler failure.

Also takes the opportunity to consolidate the print string that was occurring in each binary. The downside to this is that each project now has an "external" folder in it, all pointing to the same files. Kinda odd, since that'd probably be something worthwhlie to be put into a library and mark as a dependency, but this is purposefully trying to pull source files outside of the main source directory and include them as is.