[CI] Fix Android build on CI following update to SDK 32
What we thought was a "corrupted" installation was in fact that DX was
removed in favor of D8, and our workaround was to removed SDK 31 to revert to
version 30 that still had it.
Now we changed the workaround to a simple symlink creating from D8 to DX
for the .bat and the .jar, that will need to be removed when we find how
to make Visual Studio gradle integration properly support D8.