Kill .net 4.7.2 version of sharpmake and some related refactoring
List of changes contained in this commit:
- Enable msbuild binary log and store them as artifacts in github. Compile time is the same.
- Code signing is now disabled on macos when compiling using CI.
- Kill all mono related scripts
- Clean ifdefed code.
- No longer pass --framework argument to dotnet.exe when compiling on macos/linux. That way build is always the same on all platforms.
I think this code was different because of differences for 4.7.x on those platforms

Note: Found out that msbuild is badly handling builds where --framework is specified but only a single framework version is specified in csproj files.
It then builds twice some projects in parallel and this can cause build errors.

Removing this unifies how all platforms are handled but also workaround the msbuild bug.