Sharpmake is now using .net 6.0 instead of .net 5.0
It is about 5% faster when generating projects for one of our big game engine.

Kill extension loader code that is using domain loading and unloading:
- This is not supported anymore in .net 6
- That code was only used by ExtensionLoader.LoadExtensionsInDirectory() and it wasn't used(dead code)

Can't get a culture when running under globalization invariant in .net 6

Generate a vs2022 debug solution as this is a requirement for .net 6.

Disable Windows-2019 image when compiling on github. .net6 needs VS2022.