FastBuild: small fixes for PreBuildDependencies
Add custom file build steps for lib output

The fastBuildBuildOnlyDependencies list is only added to exe or dll outputs. If we are building a lib, make sure we add the custom file build steps to the fastBuildPreBuildDependencies list instead

Put Prebuild dependencies in the Lib/Exe/Dll section

When we have prebuild dependencies, if we build a lib, an exe or a dll, put the prebuild dependencies on that section. Since that section will include the additional object lists, the prebuild dependencies will precede all of the object lists, not just the first one on which the prebuild dependencies were previously put