[macOS] Refactor Platform and Compiler settings for Apple Platforms
This commit introduces two new global settings classes. ApplePlatformSettings
and ClangForAppleSettings. The goal is to make the compiler path and SDK path
more configurable and able to be overriden from nuget packages that contain a
custom compiler and sdk.

This introduces a breaking change in the form of deprecating
Options.Xcode.SDKRoot this has been split up per platform, meaning that if you
want to change the SDK root you need to now set ApplePlatform.Settings.MacOSSDKPath
ApplePlatform.Settings.IPhoneOSSDKPath etc.

This also introduces the ClangForApple.Settings.IsAppleClang which can be used
to check if the current compiler is the apple version of clang or a custom one.