Fix false positives with warning "Bff [path] doesn't appear in any master bff, it won't be buildable"
The warning suggests a project bff is not refered to by a master bff.
However the check was done per FastBuild conf instead of per project bff.

We allow some of the confs not to be reachable by a master bff, it can be the case for projects where a dependency is not added for all configurations.
It's not ideal but it will do for now.

 - Bff.cs checks for bffs without master Bff
 - XcodeProj.cs  checks that each bff has at most one master