Fix OrderableString resolving
Updating a value on an OrderableString using the index accessor, like if it was a simple IList, can lead to duplicated value into the underlying _list member.
And if the resolved string doesn't have the same order value than  the already existing one, it will lead to annoying error at generation time.

We fix it by using the SetOrRemoveAtIndex, which is specifically done for this kind of case